requirements.txt
requests
pyyaml
sh
